+++ lib/json_mask.rb
@@ -5,6 +5,7 @@ require_relative 'json_mask/error'
 require_relative 'json_mask/selection_tree'
 require_relative 'json_mask/parser'
 require_relative 'json_mask/projector'
+require_relative 'json_mask/exclusion_projector'
 require_relative 'json_mask/compiled_mask'
 
 # Filters JSON-compatible Ruby values using Google partial-response selectors.
@@ -41,7 +42,27 @@ module JsonMask
 
       CompiledMask.new(fields, selection_tree)
     end
+
+    def except(value, fields, **options)
+      compile_except(fields, **options).call(value)
+    end
+
+    def compile_except(
+      fields,
+      max_length: Parser::DEFAULT_MAX_LENGTH,
+      max_depth: Parser::DEFAULT_MAX_DEPTH,
+      max_selectors: Parser::DEFAULT_MAX_SELECTORS
+    )
+      selection_tree = Parser.new(
+        fields,
+        max_length:,
+        max_depth:,
+        max_selectors:
+      ).parse
+
+      CompiledMask.new(fields, selection_tree, projector: ExclusionProjector)
+    end
   end
 
-  private_constant :Parser, :Projector, :Selection, :SelectionTree
+  private_constant :Parser, :Projector, :ExclusionProjector, :Selection, :SelectionTree
 end